[feat] move test to samples. support configuration in db butterfly config to determine whether auto create required table, default is false.
