+++ butterfly-allocator/butterfly-allocator-db/src/main/java/com/simonalong/butterfly/worker/db/DbButterflyConfig.java
@@ -15,4 +15,8 @@ public class DbButterflyConfig extends ButterflyConfig {
     private String url;
     private String userName;
     private String password;
+    /**
+     * 是否自动创建相关表；true：当不存在相关表时，自动创建，false：任何情况都不自动创建
+     */
+    private Boolean autoCreateTable = false;
 }
